Add files via upload
Very simple example code connecting to the AttoDRY on COMPort 4 (check with the LabView Interface which COMPort you are using). Reads out the field and temperature, sets a field and temperature and toggles the field and temperature control. Disconnects and ends the program. Check at the AttoDRY display if the values are changed accordingly and the field and temperature control are toggled (orange instead of white on the display).